Current Update
This is the most current update, I have added a damage script and hitboxes to some of the moves. The damage script is connected to the moves and they cause the health bars to deplete when the moves attack. I still need to work on the enemy animations and movement. I also need more work in the physics department. However, the game is now reaching a stage where we just need to work on quality of life changes.